scraping_web/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
distro==1.9.0
exceptiongroup==1.3.0
groq==0.28.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
outcome==1.3.0.post0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.4
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
